Fix warning about signed/unsigned comparison in wxFile::ReadAll().
Compare ssize_t value with another ssize_t value instead of size_t one to
avoid warnings.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@72800 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/file.cpp
@@ -302,7 +302,7 @@ bool wxFile::ReadAll(wxString *str, const wxMBConv& conv)
     char* p = buf.data();
     for ( ;; )
     {
-        static const unsigned READSIZE = 4096;
+        static const ssize_t READSIZE = 4096;
 
         ssize_t nread = Read(p, length > READSIZE ? READSIZE : length);
         if ( nread == wxInvalidOffset )
